common/common_b/requirements.txt
gcloud>=0.18.3
google-api-python-client>=2.30.0
google-cloud-bigquery>=2.30.1
google-cloud-bigquery[pandas]
google-cloud-secret-manager>=2.7.0
gspread>=5.1.1
numpy>=1.17.0
oauth2client>=4.1.3
pandas>=0.25.3
pandas-gbq>=0.14.1
protobuf>=3.19.1
simplejson>=3.17.6
